[PrimTorch] Reference for matrix_norm
As per title. I corrected a thing or two from my previous implementation
to make for better errors in some weird edge-cases and have a more clear
understanding of when does this function support low_precision types and
when it doesn't.

We also use the optimisation for bfloat16 within `vector_norm` within
this function.

ghstack-source-id: 934c39db2359ab9e2f643b314c6e4774286fbc94
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81113