Lower randint default dtype to the C++ API
`randint` is currently declared with `dtype=None` but the python
binding code is special-cased to default it to `long` instead. This
leads to an inconsistency between the eager-mode python API and
everything else. So, instead I move the default handling to C++.

Note the `_window` functions also need modified here because they call
`arange` and expect it to return a float by default.

ghstack-source-id: 3774902ff9a343e67d4c6ac1ddf9e3c4188e20ba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81410