[ao] Standardize PerChannelDetector Output to be single level
Summary: Currently, the PerChannelDetector has a multi-layered output.

Example:
```
{'backend': 'qnnpack',
 'per_channel_status': {'block1.linear': {'per_channel_supported': True,
                                          'per_channel_used': False},
                        'block2.linear': {'per_channel_supported': True,
                                          'per_channel_used': False}}}
```

The issue with this is that when it comes to future features such as
visualizations where we need to go through this dictionary, it can be
hard because of the variable number of layers.

This changes the output format of the PerChannelDetector to have a
standard format.

Ex.)
```
{'block1.linear': {'backend': 'qnnpack',
                   'per_channel_supported': True,
                   'per_channel_used': False},
 'block2.linear': {'backend': 'qnnpack',
                   'per_channel_supported': True,
                   'per_channel_used': False}}
```

Test Plan: python test/test_quantization.py TestFxModelReportDetector

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 243f2628d7ef30cab11e1a632f9bec77da0d2a33
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81585