[ao] Get feature and module names from ModelReportVisualizer class
Summary: Added the functionality to be able to get the feature names and
module_fqns from the ModelReportVisualizer class. The purpose of this
addition is so that users can see the exact set of module_fqns or
feature names that they can filter based on, and use this information to
perform their filtering.

Test Plan: python test/test_quantization.py
TestFxModelReportVisualizer.test_get_modules_and_features

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: fba8cf6be5c220bf35c9f76fd1ca0a8f0b8c8980
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81647